+++ app_build_suite/build_steps/__init__.py
@@ -1,2 +1,2 @@
 """Main steps related definition and provider specific implementation."""
-from .build_step import BuildStep, BuildStepsFilteringPipeline, ALL_STEPS  # noqa: F401
+from step_exec_lib.build_step import BuildStep, BuildStepsFilteringPipeline, ALL_STEPS  # noqa: F401
