+++ app_build_suite/build_steps/repositories.py
@@ -5,8 +5,8 @@ from typing import cast
 from pykube import HTTPClient, Service
 
 from app_build_suite.build_steps import helm
-from app_build_suite.errors import TestError
-from app_build_suite.types import Context
+from step_exec_lib.errors import TestError
+from step_exec_lib.types import Context
 
 
 logger = logging.getLogger(__name__)
