+++ app_test_suite/cluster_providers/kind_cluster_provider.py
@@ -5,10 +5,11 @@ import uuid
 
 import configargparse
 
-from app_build_suite.cluster_providers import cluster_provider
-from app_build_suite.errors import TestError
-from app_build_suite.utils import files, config as config_abs
-from app_build_suite.utils.processes import run_and_log
+from app_test_suite.cluster_providers import cluster_provider
+from step_exec_lib.errors import TestError
+from step_exec_lib.utils import files
+from step_exec_lib.utils import config as config_abs
+from step_exec_lib.utils.processes import run_and_log
 
 logger = logging.getLogger(__name__)
