+++ step_exec_lib/components.py
@@ -6,8 +6,8 @@ from typing import List
 import configargparse
 
 from app_build_suite.build_steps import BuildStep
-from app_build_suite.errors import Error
-from app_build_suite.types import Context
+from step_exec_lib.errors import Error
+from step_exec_lib.types import Context
 
 logger = logging.getLogger(__name__)
